+++ hellorust/src/spawner.rs
@@ -77,7 +77,7 @@ pub fn goblin(ecs: &mut World, x: i32, y: i32) {
     monster(ecs, x, y, rltk::to_cp437('g'), "Goblin");
 }
 
-pub fn monster<S: ToString>(ecs: &mut World, x: i32, y: i32, glyph: u8, name: S) {
+pub fn monster<S: ToString>(ecs: &mut World, x: i32, y: i32, glyph: u16, name: S) {
     ecs.create_entity()
         .with(Position { x, y })
         .with(Renderable {
